fix multiple select component
